backend_chris/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy>=2.2
trimesh>=4.5
Pillow>=11
